Adjust how character recipes are exposed.
This brings it inline with how the other parts of /v2/characters are
exposed. It might seem kinda weird that it requires the "inventories"
scope, but it's been in the description for awhile -- the original
reasoning was that crafting helper tools would want the recipes (and they'd also
want inventories) so that scope was chosen to minimize the number of
grants.