Add tier_text to /v2/achievements.
This change adds `tier_text` to allow displaying of tier text. The tier text can be seen in the tooltips of achievement watchlist HUD (on the right side of the screen, "Tier: X/Z Foobar\nTotal: Y/Z Foobar") and the achievement categories view (by selecting a category, and then viewing the tooltip for an achievement).

Some achievements do not have any text for tiers (e.g. [Birthday—Year 1](https://api.guildwars2.com/v2/achievements/1681)). This can be represented as a blank string, or by making this field optional.

For one-argument strings (i.e. `"%str1% Dungeon Story[pl:\"Stories\"] Complete"`), singular form is followed by plural form (e.g. `Item[s]` or `Match[pl:"Matches"]`). This is a bonus to find grammatical errors in the game data — namely single argument strings where a tier's count is 1 but the text is plural, such as "1 Ranked Matches Won as a Mesmer" (typo is only as an example; game uses correct singular form for this achievement)