Break /v2/commerce/listings to not return the entire set.
Currently /v2/commerce/listings returns potentially huge responses which
are only useful for a small subset of applications -- but a small subset
of the data returned is valuable more widely. This incongruity is causing
some undue stress; the goal of this breaking change is to limit the
response size for the common case while still providing access to the full
dataset for applications which need it.

By applying pagination.