/v2/backstory/* -- make answer ids strings.
Apparently answer ids aren't guarenteed to be unique -- the mesmer and
revenant backstory question has the same answer ids (that's the only case
of overlap). To support this, change the answer ids to be the
concatenation of the parent question with the original answer id (so that
mesmer/revenant backstory answers have unique ids).

This is an unfortunate change and is very backwards incompatible, but I
couldn't come up with a better solution :(