First-pass at some bulk-expanded story endpoints.
Right now the chapters section is a bit bare -- there's not much metadata
the backend service currently exposes for these. Will probably add the
chapter rewards and locations (and maybe event metadata, if I have the
courage to dig that out) as part of a future pull request.

This also kindly omits the account side of things for exposing which
chapters have been completed (which is on a per-character basis). This is
a bit more finicky because the personal story is branching -- not entirely
certain how to best express that.

Finally, on a related note, the relations between chapters is currently
pretty vague (e.g., that only one of the orders can be selected). Still
need to figure out how to express that in a coherent way.